+++ pkg/config/config.go
@@ -1,8 +1,15 @@
 package config
 
 import (
+	"log/slog"
 	"os"
 	"path/filepath"
+)
+
+var (
+	outputCompression = "zip"
+	outputCSVType     = "tsv"
+	jobsNum           = 5
 )
 
 // Config is a configuration object for the Darwin Core Archive (DwCA)
@@ -22,6 +29,13 @@ type Config struct {
 	// data.
 	OutputPath string
 
+	// OutputArchiveCompression is the compression format to use when
+	// creating the output archive. It can be "zip" or "tar.gz".
+	OutputArchiveCompression string
+
+	// OutputCSVType is the type of CSV files. Can be "csv" or "tsv"
+	OutputCSVType string
+
 	// JobsNum is the number of concurrent jobs to run.
 	JobsNum int
 }
@@ -30,10 +44,53 @@ type Config struct {
 // the configuration are organized.
 type Option func(*Config)
 
-// OptPath sets the root path for all temporary files.
-func OptPath(s string) Option {
+// OptRootPath sets the root path for all temporary files.
+func OptRootPath(s string) Option {
 	return func(c *Config) {
 		c.RootPath = s
+	}
+}
+
+// OptOutputArchiveCompression sets the compression format to use when
+// creating the output archive. It can be "zip" or "tar.gz".
+func OptArchiveCompression(s string) Option {
+	return func(c *Config) {
+		if s != "zip" && s != "tar" {
+			slog.Warn(
+				"Entered compression format is not supported. Using default format",
+				"input", s, "default", outputCompression,
+			)
+			s = outputCompression
+		}
+		c.OutputArchiveCompression = s
+	}
+}
+
+// OptOutputCSVType sets the type of CSV files. Can be "csv" or "tsv"
+func OptOutputCSVType(s string) Option {
+	return func(c *Config) {
+		if s != "csv" && s != "tsv" {
+			slog.Warn(
+				"Entered CSV type is not supported. Using default format",
+				"bad-input", s, "default", outputCSVType,
+			)
+			s = outputCSVType
+		}
+		c.OutputCSVType = s
+	}
+}
+
+// OptJobsNum sets the number of concurrent jobs to run.
+func OptJobsNum(n int) Option {
+	return func(c *Config) {
+		if n < 1 || n > 100 {
+			slog.Warn(
+				"Unsupported number of jobs (supported: 1-100). Using default value",
+				"bad-input", n, "default", jobsNum,
+			)
+			n = jobsNum
+		}
+		c.JobsNum = n
 	}
 }
 
@@ -47,8 +104,10 @@ func New(opts ...Option) Config {
 
 	path = filepath.Join(path, "dwca_go")
 	c := Config{
-		RootPath: path,
-		JobsNum:  5,
+		RootPath:                 path,
+		OutputArchiveCompression: outputCompression,
+		OutputCSVType:            outputCSVType,
+		JobsNum:                  jobsNum,
 	}
 
 	for _, opt := range opts {